# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vowed API",
    "version": "0.1.0",
    "description": "Backend for the Vowed Android app. All endpoints except /v1/health, /v1/meta and /v1/auth/* need `Authorization: Bearer <token>` from /v1/auth/verify. u64 amounts are decimal strings. Transaction builders return unsigned base64 transactions: the app must decode and check them before asking the wallet to sign."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/devices/register": {
      "post": {
        "summary": "Register the device proof key (P-256) with its Android key attestation",
        "tags": [
          "devices"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "devicePublicKey": {
                    "type": "string",
                    "maxLength": 400
                  },
                  "attestationChain": {
                    "maxItems": 6,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "maxLength": 6000
                    }
                  },
                  "challenge": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "walletSignature": {
                    "type": "string",
                    "maxLength": 200
                  }
                },
                "required": [
                  "devicePublicKey",
                  "challenge",
                  "walletSignature"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deviceId": {
                      "type": "string"
                    },
                    "trustCap": {
                      "type": "string",
                      "enum": [
                        "high",
                        "medium",
                        "low"
                      ]
                    },
                    "attestation": {
                      "type": "object",
                      "properties": {
                        "level": {
                          "type": "string"
                        },
                        "verifiedBoot": {
                          "type": "string"
                        },
                        "note": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "level",
                        "verifiedBoot",
                        "note"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "deviceId",
                    "trustCap",
                    "attestation"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/proofs/session": {
      "post": {
        "summary": "Start a proof session: returns the nonce the device must sign into the proof",
        "tags": [
          "proofs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pool": {
                    "type": "string"
                  },
                  "dayIndex": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 59
                  },
                  "proofType": {
                    "type": "string",
                    "enum": [
                      "CAMERA_POSE",
                      "USAGE_LIMIT",
                      "NO_USE_WINDOW",
                      "STEPS",
                      "GEOFENCE",
                      "FOCUS_TIMER",
                      "SELF_ATTEST"
                    ]
                  }
                },
                "required": [
                  "pool",
                  "dayIndex",
                  "proofType"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": {
                      "type": "string"
                    },
                    "nonce": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "number"
                    },
                    "proofType": {
                      "type": "string",
                      "enum": [
                        "CAMERA_POSE",
                        "USAGE_LIMIT",
                        "NO_USE_WINDOW",
                        "STEPS",
                        "GEOFENCE",
                        "FOCUS_TIMER",
                        "SELF_ATTEST"
                      ]
                    },
                    "dayIndex": {
                      "type": "number"
                    },
                    "target": {
                      "type": "object",
                      "properties": {
                        "metric": {
                          "type": "string"
                        },
                        "value": {
                          "type": "number"
                        },
                        "unit": {
                          "type": "string"
                        },
                        "direction": {
                          "type": "string",
                          "enum": [
                            "atLeast",
                            "atMost"
                          ]
                        }
                      },
                      "required": [
                        "metric",
                        "value",
                        "unit",
                        "direction"
                      ],
                      "additionalProperties": false
                    },
                    "window": {
                      "type": "object",
                      "properties": {
                        "opensAt": {
                          "type": "number"
                        },
                        "closesAt": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "opensAt",
                        "closesAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "sessionId",
                    "nonce",
                    "expiresAt",
                    "proofType",
                    "dayIndex",
                    "target",
                    "window"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/proofs/submit": {
      "post": {
        "summary": "Submit a device-signed proof package. Idempotent per session: a retry returns the same result",
        "tags": [
          "proofs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sessionId": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 64
                  },
                  "nonce": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 128
                  },
                  "challengeId": {
                    "type": "string",
                    "minLength": 32,
                    "maxLength": 44
                  },
                  "dayIndex": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 59
                  },
                  "proofType": {
                    "type": "string",
                    "enum": [
                      "CAMERA_POSE",
                      "USAGE_LIMIT",
                      "NO_USE_WINDOW",
                      "STEPS",
                      "GEOFENCE",
                      "FOCUS_TIMER",
                      "SELF_ATTEST"
                    ]
                  },
                  "metrics": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "startedAt": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991
                  },
                  "endedAt": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991
                  },
                  "evidenceHash": {
                    "type": "string",
                    "pattern": "^[0-9a-f]{64}$"
                  },
                  "deviceKeyId": {
                    "type": "string",
                    "pattern": "^[0-9a-f]{64}$"
                  },
                  "signature": {
                    "type": "string",
                    "minLength": 80,
                    "maxLength": 200
                  }
                },
                "required": [
                  "sessionId",
                  "nonce",
                  "challengeId",
                  "dayIndex",
                  "proofType",
                  "metrics",
                  "startedAt",
                  "endedAt",
                  "evidenceHash",
                  "deviceKeyId",
                  "signature"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accepted": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "trustTier": {
                      "type": "string",
                      "enum": [
                        "high",
                        "medium",
                        "low"
                      ]
                    },
                    "daysCompleted": {
                      "type": "number"
                    },
                    "streak": {
                      "type": "number"
                    },
                    "checkin": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "confirmed",
                            "pending",
                            "failed"
                          ]
                        },
                        "signature": {
                          "type": "string"
                        },
                        "error": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "status"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "accepted",
                    "trustTier",
                    "daysCompleted",
                    "streak",
                    "checkin"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}